Transitioned to data_block for parsing payload data. Sends all test data over the serial_test python module generating test data as payload messages.
