+++ modules/serial/serial_test.py
@@ -1,9 +1,13 @@
 import multiprocessing
 import random
+import struct
 import time
 from multiprocessing import Process, Queue
 import serial
 
+altitude = 0
+temp = 22
+going_up = True
 
 class SerialTestClass(Process):
     def __init__(self, serial_input: Queue, serial_output: multiprocessing.Queue, COM_PORT="COM1"):
@@ -12,6 +16,13 @@ class SerialTestClass(Process):
         self.SERIAL_PORT = COM_PORT
         self.serial_input = serial_input
         self.serial_output = serial_output
+
+
+        # Continually add to these
+        self.altitude = altitude
+        self.temp = temp
+        self.going_up = going_up
+
         self.run()
 
     def run(self):
@@ -27,35 +38,40 @@ class SerialTestClass(Process):
 
         call_sign = b"Devils".hex()
         # print("Call sign -> ", call_sign)
-        self.serial_output.put(payload)
+        #self.serial_output.put(payload)
 
         packet_callsign = "446576696c73"
         packet_callsignX = "446576696c73202020202020"
 
 
-        packet_call_sign = b"Devils".hex()
-        six_byte_spacer = b"      ".hex()
-        block_header = "840C0000"
-        block_altitude_rocket_on_fire = "E01F00008D540100BC57FF0010FEFFFF"
-        block_altitude =                "E01F00008D540100BC44010020FF0600"
-        block_altitude_zero =           "E01F00008D540100BC44010000000000"
-        block_altitude_zero_fire =           "E01F00008D540100BC57FF0000000000"
-        #time.sleep(5)
-        print(f"BEST {packet_call_sign}{six_byte_spacer}{block_header}{block_altitude}")
-
-        self.serial_output.put(f"{packet_call_sign}{six_byte_spacer}{block_header}{block_altitude_rocket_on_fire}")
-        #time.sleep(1)
-        self.serial_output.put(f"{packet_callsign}{six_byte_spacer}{block_header}{block_altitude_zero}")
-        time.sleep(5)
-        self.serial_output.put(f"{packet_callsignX}{block_header}{block_altitude_zero_fire}")
-        time.sleep(2)
-        self.serial_output.put(f"{packet_call_sign}{six_byte_spacer}{block_header}{block_altitude}")
-
-        #self.serial_output.put(f"{packet_call_sign}{six_byte_spacer}")
         while True:
             random_data = int(random.uniform(0, 1000))
             #print(f"SERIAL TEST: {random_data}")
             #self.serial_output.put(random_data)
             #print("Read from serial", test_serial.readline())
-            time.sleep(.2)
+            time.sleep(1)
             #print("Read from serial b''")
+            self.tester()
+
+    def tester(self):
+        random_alternation = int(random.uniform(0, 1000))
+        if self.going_up:
+            self.temp += random_alternation / 500
+        else:
+            self.temp -= random_alternation / 500
+
+        if self.temp > 100:
+            self.going_up = False
+        elif self.temp < 20:
+            self.going_up = True
+
+        self.altitude += random.uniform(0, 3)
+
+
+        packet_call_sign = b"Devils".hex()
+        six_byte_spacer = b"      ".hex() # Should be packet header data!!!
+        packet_header = f"{packet_call_sign}{six_byte_spacer}"
+        block_header = "840C0000"  # Should be struct generated header data!!!
+
+        #self.serial_output.put(f"{packet_call_sign}{six_byte_spacer}{block_header}{'E01F00008D540100BC57FF0010FEFFFF'}")
+        self.serial_output.put(f"{packet_header}{block_header}{struct.pack('<Iiii', 8160, 87181, int(self.temp*1000), int(self.altitude*1000)).hex().upper()}")
