+++ modules/telemetry/telemetry.py
@@ -1,40 +1,8 @@
 import multiprocessing
-import random
 import struct
 import time
 
-from modules.telemetry import packets
-from modules.misc.converter import celsius_to_fahrenheit, metres_to_feet
-
-# Constants
-SUBTYPE = {
-    0: "debug_message",
-    1: "status",
-    2: "startup_message",
-    3: "altitude_data",
-    4: "acceleration_data",
-    5: "angular_velocity_data",
-    6: "GNSS_location_data",
-    7: "GNSS_metadata",
-    8: "power_information",
-    9: "temperatures",
-    "A": "MPU9250_IMU_data",
-    "B": "KX134-1211_accelerometer_data"
-}
-
-PACKETS = {
-    SUBTYPE[3]: packets.AltitudeData,
-    SUBTYPE[4]: packets.AccelerationData,
-    SUBTYPE[5]: packets.AngularVelocityData,
-    SUBTYPE[6]: packets.GNSSLocationData,
-    SUBTYPE[7]: packets.GNSSMetaData,
-    SUBTYPE["A"]: packets.MPU9250Data,
-    SUBTYPE["B"]: packets.KX1341211Data
-}
-
-altitude = 0
-temp = 22
-going_up = True
+from modules.telemetry.data_block import DataBlock, DataBlockSubtype
 
 
 class Telemetry(multiprocessing.Process):
@@ -42,16 +10,11 @@ class Telemetry(multiprocessing.Process):
                  telemetry_json_output: multiprocessing.Queue):
         super().__init__()
 
-        # Continually add to these
-        self.altitude = altitude
-        self.temp = temp
-        self.going_up = going_up
-
         self.serial_input = serial_input
         self.serial_data_output = serial_data_output
         self.telemetry_json_output = telemetry_json_output
 
-        # Telemetry Data holds a dictionary of the latest copy of each received data block stored under the SUBTYPE key.
+        # Telemetry Data holds a dict of the latest copy of received data blocks stored under the subtype name as a key.
         self.telemetry_data = {}
         self.status_data = {
             "board": {
@@ -73,79 +36,8 @@ class Telemetry(multiprocessing.Process):
                 self.parse_rx(self.serial_data_output.get())
 
             # print(f"Telemetry sending sample websocket json")
-            self.send_sample_websocket_json()
             time.sleep(.100)
-            #self.telemetry_json_output.put(self.generate_websocket_response())
-
-    def send_sample_websocket_json(self):
-        random_alternation = int(random.uniform(0, 1000))
-
-        if self.going_up:
-            self.temp += random_alternation / 500
-        else:
-            self.temp -= random_alternation / 500
-
-        if self.temp > 100:
-            self.going_up = False
-        elif self.temp < 20:
-            self.going_up = True
-
-        self.altitude += random.uniform(0, 3)
-        metres = self.altitude
-
-        sample_json = {
-            "version": "0.1.1",
-            "org": "CU InSpace",
-            "status": {
-                "board": {
-                    "connected": "yes"
-                },
-                "rocket": {
-                    "call_sign": "Not a missile",
-                    "status_code": 2,
-                    "status_name": "POWERED ASCENT",
-                    "last_mission_time": 8120
-                }
-            },
-            "telemetry_data": {
-                "altitude_data": {
-                    "mission_time": "8120",
-                    "pressure": {
-                        "pascals": "87181"
-                    },
-                    "altitude": {
-                        "metres": f"{metres}",
-                        "feet": f"{metres_to_feet(metres)}"
-                    },
-                    "temperature": {
-                        "celsius": f"{round(self.temp, 3)}",
-                        "fahrenheit": f"{celsius_to_fahrenheit(self.temp)}"
-                    }
-                },
-                "acceleration_data": {
-                    "mission_time": "8120",
-                    "acceleration": {
-                        "ms2": "40",
-                        "fts2": "131"
-                    }
-                },
-                "position_data": {
-                    "mission_time": "8120",
-                    "coordinates": {
-                        "lat": "45.385273194259014",
-                        "long": "-75.69838116342798"
-                    },
-                    "gnss_metadata": {
-                        "gnss_sat_type": "gps",
-                        "gps_sats_in_view": 3,
-                        "glonass_sats_in_view": 2,
-                        "sats_in_view": 5
-                    }
-                }
-            }
-        }
-
-        self.telemetry_json_output.put(sample_json)
+            self.telemetry_json_output.put(self.generate_websocket_response())
 
     def generate_websocket_response(self, telemetry_keys="all"):
         return {"version": "0.2.3", "org": "CU InSpace", "status": self.generate_status_data(),
@@ -177,8 +69,7 @@ class Telemetry(multiprocessing.Process):
         blocks = data[24:]  # Remove the packet header
 
         print("-----" * 20)
-        print(f'{bytes.fromhex(call_sign)} sent you a packet:')
-
+        print(f'{str(bytes.fromhex(call_sign))} sent you a packet:')
         # Parse through all blocks
         # TODO Catch type&subtype 0 and do a signal report.
         #  self.serial_input.put('radio get snr')
@@ -189,21 +80,18 @@ class Telemetry(multiprocessing.Process):
         while blocks != '':
             # Parse block header
             block_header = blocks[:8]
-            block_len, crypto_signature, message_type, message_subtype, dest_addr = _parse_block_header(block_header)
+            block_len, crypto_signature, block_type, block_subtype, dest_addr = _parse_block_header(block_header)
 
             block_len = block_len * 2  # Convert length in bytes to length in hex symbols
-            payload = blocks[8: 8 + block_len]
+            payload = bytes.fromhex(blocks[8: 8 + block_len])
 
-            # Create data if correct packet format was found
-            packet = PACKETS.get(SUBTYPE.get(message_subtype))
-            packet_data = packet.create_from_raw(payload) if packet else None
-            print(f"BLOCK DATA: {payload}")
-            print(f"BLOCK FOR {SUBTYPE.get(message_subtype)} RETURNED {dict(packet_data)}")
+            block_contents = DataBlock.from_payload(DataBlockSubtype(block_subtype), payload)
+            self.telemetry_data[DataBlockSubtype(block_subtype).name.lower()] = dict(block_contents)
+            print(block_contents)
 
-            #print("BLOCKHEADER:", _parse_block_header("840C0000"))
-            #print("BLOCKHEADER:", _parse_block_header("8400B000"))
-            #print("BLOCKHEADER:", _parse_block_header("840AD000"))
-            self.telemetry_data[SUBTYPE.get(message_subtype)] = dict(packet_data)
+            # print("BLOCKHEADER:", _parse_block_header("840C0000"))
+            # print("BLOCKHEADER:", _parse_block_header("8400B000"))
+            # print("BLOCKHEADER:", _parse_block_header("840AD000"))
 
             # Remove the data we processed from the whole set, and move onto the next data block
             blocks = blocks[8 + block_len:]
@@ -242,27 +130,12 @@ def _parse_block_header(header) -> tuple:
     block_len: int
     crypto_signature: bool
     message_type: int
-    message_subtype: str
+    message_subtype: int
     destination_addr: int
     """
-    #header = OG_Header
-    #header = bin(int(header, 16))  # Convert into binary
-
-
-    #block_len: int = (int(header[0:5], 2) + 1) * 4  # Length of block in bytes
-    #crypto_signature: bool = True if int(header[5], 2) == 1 else False  # Check if message has a cryptographic signature
-
-    # Type of message (Ex: The purpose of the block, such as transmitting info to ground station)
-    #message_type: int = int(header[6:10], 2)
-    #print("RAW HEADER[6:10]",header[6:10])
-
-    #message_subtype: str = SUBTYPE.get(int(header[10:16], 2))  # Type of information that is stored in the block
-    #destination_addr: int = int(header[16:20], 2)
-    #print("PARSEHEADER", block_len, crypto_signature, message_type, message_subtype, destination_addr)
-
     header = struct.unpack('<I', bytes.fromhex(header))
 
-    block_len = ((header[0] & 0x1f) + 1) * 4
+    block_len = ((header[0] & 0x1f) + 1) * 4  # Length of the data block
     crypto_signature = ((header[0] >> 5) & 0x1)
     message_type = ((header[0] >> 6) & 0xf)  # 0 - Control, 1 - Command, 2 - Data
     message_subtype = ((header[0] >> 10) & 0x3f)
